Style updates, comment out mx deck so no chance of it showing up in recording
